+++ src/main/java/kr/co/khacademy/semi/controller/admin/AdminUserController.java
@@ -1,6 +1,5 @@
-package kr.co.khacademy.semi.controller;
+package kr.co.khacademy.semi.controller.admin;
 
-import kr.co.khacademy.semi.dao.AccountDao;
 import kr.co.khacademy.semi.dao.UserDao;
 import kr.co.khacademy.semi.model.User;
 
@@ -12,11 +11,10 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
-@WebServlet("/admin/account/normal-user/*")
-public class NormalUserController extends HttpServlet {
+@WebServlet("/admin/account/admin-user/*")
+public class AdminUserController extends HttpServlet {
 
     private static final UserDao userDao = UserDao.getInstance();
 
@@ -25,20 +23,20 @@ public class NormalUserController extends HttpServlet {
         try {
             String pathInfo = req.getPathInfo();
             if ("/list".equals(pathInfo)) {
-                List<User> normalUsers = userDao.read();
-                normalUsers.stream()
+                List<User> adminUsers = userDao.read();
+                adminUsers.stream()
                     .filter((user) -> {
-                        return user.getAccount().getRoleId() == 1;
+                        return user.getAccount().getRoleId() != 1;
                     })
                     .collect(Collectors.toList()
                 );
-                req.setAttribute("normalUsers", normalUsers);
-                req.getRequestDispatcher("/WEB-INF/views/admin/account/normal_user/list.jsp").forward(req, resp);
+                req.setAttribute("adminUsers", adminUsers);
+                req.getRequestDispatcher("/WEB-INF/views/admin/account/admin_user/list.jsp");
             } else if ("/item".equals(pathInfo)) {
                 Long id = Long.valueOf(req.getParameter("id"));
-                User normalUser = userDao.read(id);
-                req.setAttribute("normalUser", normalUser);
-                req.getRequestDispatcher("/WEB-INF/views/admin/account/normal_user/item.jsp").forward(req, resp);
+                User adminUser = userDao.read(id);
+                req.setAttribute("adminUser", adminUser);
+                req.getRequestDispatcher("/WEB-INF/views/admin/account/admin_user/list.jsp");
             }
         } catch (SQLException e) {
             resp.sendRedirect("/error.jsp");
@@ -49,7 +47,11 @@ public class NormalUserController extends HttpServlet {
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         try {
             String pathInfo = req.getPathInfo();
-            if ("/modify".equals(pathInfo)) {
+            if ("/register".equals(pathInfo)) {
+                User user = User.of(req);
+                userDao.create(user);
+                resp.sendRedirect("/admin/account/admin-user/list");
+            } else if ("/modify".equals(pathInfo)) {
                 User user = User.of(req);
                 Long targetId = Long.valueOf(req.getParameter("target-id"));
                 userDao.update(user, targetId);
